Moved files to ads folder
