+++ js/adapters/Bitmovin8Adapter.ts
@@ -5,7 +5,7 @@ import {Adapter} from '../types/Adapter';
 import {AdapterEventCallback} from '../types/AdapterEventCallback';
 import {DrmPerformanceInfo} from '../types/DrmPerformanceInfo';
 import {AdClickedEvent, AdQuartileEvent, ErrorEvent, AdEvent, AdBreakEvent} from 'bitmovin-player';
-import { AdCallbacks } from '../types/AdCallbacks';
+import { AdCallbacks } from '../types/ads/AdCallbacks';
 
 class Bitmovin8Adapter implements Adapter {
   onBeforeUnLoadEvent: boolean;
