+++ js/enums/ads/StrategyType.ts
