+++ js/types/ads/AdCallbacks.ts
