+++ js/types/ads/AnalyticsAdBreak.ts
@@ -1,0 +1,7 @@
+import { AdSample } from './AdSample';
+
+export interface AnalyticsAdBreak {
+  id: string;
+  adStarted: number;
+  adSample: Array<AdSample>
+}
